Move string formatting out of "except" files.
