+++ core/include/scipp/core/variable.h
@@ -16,6 +16,7 @@
 #include "scipp/core/dimensions.h"
 #include "scipp/core/dtype.h"
 #include "scipp/core/slice.h"
+#include "scipp/core/string.h"
 #include "scipp/core/variable_view.h"
 #include "scipp/core/vector.h"
 #include "scipp/units/unit.h"
@@ -816,14 +817,6 @@ SCIPP_CORE_EXPORT Variable atan(const Variable &var);
 
 template <class T>
 VariableView<const T> getView(const Variable &var, const Dimensions &dims);
-
-SCIPP_CORE_EXPORT std::ostream &operator<<(std::ostream &os,
-                                           const VariableConstProxy &variable);
-SCIPP_CORE_EXPORT std::ostream &operator<<(std::ostream &os,
-                                           const VariableProxy &variable);
-SCIPP_CORE_EXPORT std::ostream &operator<<(std::ostream &os,
-                                           const Variable &variable);
-SCIPP_CORE_EXPORT std::ostream &operator<<(std::ostream &os, const Dim dim);
 
 } // namespace scipp::core
